+++ src/index.tsx
@@ -1,3 +1,6 @@
 import DatePicker from './date-picker';
+import './styles.css';
+
+export { default as TimePicker } from './time-picker';
 
 export default DatePicker;
